provigen sqlcipher support goal: support either SQLite or SqlCipher.
design: abstract the common behavior into the interfaces Database and OpenHelper and implementations of
these. for SQLite the library uses AndroidSQLiteDatabase and ProviGenSQLiteOpenHelper. for SqlCipher it
uses SQLCipherDatabase and ProviGenSQLCipherOpenHelper.

use: an application using SQLite should use ProviGenSQLiteOpenHelper in its ContentProvider in place of
SQLiteOpenHelper. for SqlCipher, the application uses ProviGenSQLCipherOpenHelper. SqlCipher also requires
the app to make to calls:
	SQLiteDatabase.loadLibs(getApplicationContext());
	ProviGenSQLCipherOpenHelper.setPassword("swordfish");

these are usually called in Application.onCreate(). they should be called before any use of the database.
loadLibs() is required for any app using SqlCipher. as the SqlCipher database file is encrypted call
ProviGenSQLCipherOpenHelper.setPassword() to provide the password used to decrypt it.

a new sample, ProviGenSqlcipherSample, shows how to use  SqlCipher with ProviGen.